fix: frontend refactoring /courses
- Рефакторинг с учетом изменения структуры данных сущности Course
(смена PK from uuid to string)
- Перенос CoursesLayout.stub.tsx в \pages\courses\list-stub\
